+++ web/src/app/actions.ts
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation';
+
+export const goTo = (route: string) => {
+    redirect(route);
+};
