+++ web/src/app/layout.tsx
@@ -1,7 +1,8 @@
+'use client';
+
 import { ColorSchemeScript } from '@mantine/core';
 import '@mantine/core/styles.css';
 import { Amplify } from 'aws-amplify';
-import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { PropsWithChildren } from 'react';
 import amplifyConfig from '../config/amplifyConfig';
@@ -10,13 +11,10 @@ import Providers from './providers';
 
 const inter = Inter({ subsets: ['latin'] });
 
-export const metadata: Metadata = {
-    title: 'In-Ctrl',
-    description:
-        'InCtrl is an open-source version control system built for ultimate quality control. "Git" in control with InCtrl ',
-};
-
-Amplify.configure(amplifyConfig);
+console.log(amplifyConfig);
+Amplify.configure(amplifyConfig, {
+    ssr: true,
+});
 
 const RootLayout = ({ children }: PropsWithChildren) => {
     return (
