+++ web/src/app/page.tsx
@@ -1,5 +1,12 @@
 import { Routes } from '@/models/routes/Routes';
+import { Metadata } from 'next';
 import Link from 'next/link';
+
+export const metadata: Metadata = {
+    title: 'In-Ctrl',
+    description:
+        'InCtrl is an open-source version control system built for ultimate quality control. "Git" in control with InCtrl ',
+};
 
 const Home = () => {
     return (
